fix: move utils into src
